### 2018-01-30 #560 @whedon
[PRE REVIEW]: PyMap3D: 3-D coordinate conversions for terrestrial and geospace environments
**Submitting author:** @scivision (<a href="http://orcid.org/0000-0002-1637-6526">Michael Hirsch</a>)
**Repository:** <a href="https://github.com/scivision/pymap3d" target ="_blank">https://github.com/scivision/pymap3d</a>
**Version:** v1.5.0
