+++ paper/paper.bib
@@ -6,7 +6,7 @@
   journal = {Survey Review},
   volume = 23,
   number = 176,
-  year = 1975
+  year = 1975,
   month = 4,
   pages = {88--93},
   url = {https://www.ngs.noaa.gov/PUBS_LIB/inverse.pdf},
