### 2018-01-31 #560 @lheagy
Hi, @arfon, this looks identical to #555, did something go awry here?
