### 2018-01-31 #560 @scivision
Hi Lindsey, I withdrew and resubmitted as I submitted an old release--I resubmitted with the current pymap3d release.
