### 2018-01-31 #560 @lheagy
:wave: Hi @jsta, @nuest, @fgnievinski, @hugoledoux, @sgrieve, @fil, would any of you be willing to review this submission for JOSS?
