+++ pymap3d/__init__.py
@@ -18,8 +18,13 @@ see tests/Test.py for example uses.
 from __future__ import division
 from six import string_types,PY2
 from datetime import datetime
-import numpy as np
-from numpy import sin, cos, tan, sqrt, radians, arctan2, hypot, degrees
+try:
+    import numpy
+    from numpy import sin, cos, tan, sqrt, radians, arctan2, hypot, degrees
+except ImportError:
+    numpy = None
+    from math import sin, cos, tan, sqrt, radians, hypot, degrees
+    from math import atan2 as arctan2
 try:
     from astropy.time import Time
     from astropy import units as u
@@ -136,10 +141,10 @@ def eci2ecef(eci, t):
 
     output: ECEF x,y,z (meters)
     """
-    if Time is None:
-        raise ImportError('You need to install AstroPy')
+    if numpy is None or Time is None:
+        raise ImportError('eci2ecef requires Numpy and AstroPy')
 
-    t = np.atleast_1d(t)
+    t = numpy.atleast_1d(t)
     if isinstance(t[0], string_types): #don't just ram in in case it's float
         t = str2dt(t)
 
@@ -152,14 +157,14 @@ def eci2ecef(eci, t):
 
     assert isinstance(gst[0], float)  # must be in radians!
 
-    eci = np.atleast_2d(eci)
+    eci = numpy.atleast_2d(eci)
     N, trip = eci.shape
     if eci.ndim > 2 or trip != 3:
         raise ValueError('eci triplets must be shape (N,3)')
     """ported from:
     https://github.com/dinkelk/astrodynamics/blob/master/rot3.m
     """
-    ecef = np.empty_like(eci)
+    ecef = numpy.empty_like(eci)
 
     for i in range(N):
         #ecef[i, :] = _rottrip(gst[i]) @ eci[i, :]
@@ -217,9 +222,11 @@ def ned2ecef(n, e, d, lat0, lon0, h0, ell=None, deg=True):
 #%% to ECI
 def aer2eci(az, el, srange, lat0, lon0, h0, t, ell=None, deg=True):
     """convert target azimuth, elevation, range to ECI"""
+    if numpy is None:
+        raise ImportError('aer2eci requires Numpy')
     x, y, z = aer2ecef(az, el, srange, lat0, lon0, h0, ell, deg)
 
-    return ecef2eci(np.column_stack((x, y, z)), t)
+    return ecef2eci(numpy.column_stack((x, y, z)), t)
 
 
 def ecef2eci(ecef, t):
@@ -227,10 +234,10 @@ def ecef2eci(ecef, t):
     input t is either a datetime or float in radians
 
     """
-    if Time is None:
-        raise ImportError('Please install AstroPy')
+    if Time is None or numpy is None:
+        raise ImportError('ecef2eci requires Numpy and AstroPy')
 
-    t = np.atleast_1d(t)
+    t = numpy.atleast_1d(t)
     if isinstance(t[0], string_types): #don't just ram in in case it's float
         t = str2dt(t)
 
@@ -243,14 +250,14 @@ def ecef2eci(ecef, t):
 
     assert isinstance(gst[0], float)  # must be in radians!
 
-    ecef = np.atleast_2d(ecef)
+    ecef = numpy.atleast_2d(ecef)
     N, trip = ecef.shape
     if ecef.ndim > 2 or trip != 3:
         raise TypeError('ecef triplets must be shape (N,3)')
     """ported from:
     https://github.com/dinkelk/astrodynamics/blob/master/rot3.m
     """
-    eci = np.empty_like(ecef)
[…]
