+++ tests/Test.py
@@ -9,87 +9,31 @@
 """
 runs tests
 """
+import logging
 import subprocess
+import unittest
 from datetime import datetime
-from numpy import asarray,radians
-from numpy.testing import assert_allclose, assert_almost_equal,run_module_suite
+try:
+    import numpy
+    from numpy import radians
+    from pymap3d.vincenty import vreckon,vdist
+except ImportError:
+    numpy = None
+    from math import radians
 from pytz import UTC
 #
 import pymap3d as pm
 from pymap3d.haversine import angledist,angledist_meeus
-from pymap3d.vincenty import vreckon,vdist
 from pymap3d.datetime2hourangle import datetime2sidereal
 from pymap3d.vallado import vazel2radec, vradec2azel
 from pymap3d.timeconv import str2dt
 
-
-def test_bsr():
-    try:
-        from pathlib import Path
-        path = Path(__file__).parents[1]
-        subprocess.check_call(['octave-cli','-q','Test.m'], cwd=path/'tests')
-    except ImportError:
-        pass
-
-
-def test_str2dt():
-
-    assert str2dt('2014-04-06T08:00:00Z') == datetime(2014,4,6,8, tzinfo=UTC)
-    ti = [str2dt('2014-04-06T08:00:00Z'), str2dt('2014-04-06T08:01:02Z')]
-    to = [datetime(2014,4,6,8, tzinfo=UTC),  datetime(2014,4,6,8,1,2, tzinfo=UTC)]
-    assert  ti == to   # even though ti is numpy array of datetime and to is list of datetime
-
-# %%
 t = '2014-04-06T08:00:00Z'
 lat, lon = (65, -148)
 ra, dec = (166.5032081149338,55.000011165405752)
 ha = 45.482789587392013
 azi, eli = (180.1,80)
 
-def test_datetime2sidereal():
-    sra = 2.90658
-    # http://www.jgiesen.de/astro/astroJS/siderealClock/
-    sdrapy = datetime2sidereal(t, radians(lon), False)
-    assert_allclose(sdrapy, sra, rtol=1e-5)
-
-    sdrvallado = datetime2sidereal(t, radians(lon), True)
-    assert_allclose(sdrvallado, sra, rtol=1e-5)
-
-
-def test_azel2radec():
-    R,D = pm.azel2radec(azi, eli, lat, lon, t)
-    assert_allclose(R, ra, rtol=1e-2)
-    assert_allclose(D, dec, rtol=1e-2)
-
-    Rv, Dv = vazel2radec(azi, eli, lat, lon, t)
-    assert_allclose(Rv, ra)
-    assert_allclose(Dv, dec)
-
-
-def test_radec2azel():
-    azapy, elapy = pm.radec2azel(ra,dec,  lat, lon, t)
-    assert_allclose(azapy, azi, rtol=1e-2)
-    assert_allclose(elapy, eli, rtol=1e-2)
-
-    azvallado, elvallado = vradec2azel(ra, dec, lat, lon, t)
-    assert_allclose(azvallado, azi, rtol=1e-2)
-    assert_allclose(elvallado, eli, rtol=1e-2)
-
-
-def test_haversine():
-    assert_almost_equal(angledist(35,23, 84,20), ha)
-    #%% compare with astropy
-    assert_almost_equal(ha, angledist_meeus(35,23, 84,20))
-
-
-def test_vincenty():
-    az = 38; sr = 3e3
-    lat2,lon2,a21 = vreckon(10,20,sr,az)
-    assert_almost_equal((lat2,lon2,a21),
[…]
