### 2018-02-01 bda42e0
increase test coverage
