### 2018-02-01 cbb1ace
increase test coverage
