+++ .travis.yml
@@ -2,7 +2,11 @@ language: python
 fast_finish: true
 
 python:
+  - 2.6
   - 2.7
+  - 3.4
+  - 3.5
+  - 3.6
   - "3.7-dev"
   
 
