### 2018-02-01 e2b6601
test Matlab code once
