+++ .travis.yml
@@ -4,6 +4,8 @@ fast_finish: true
 python:
   - 2.6
   - 2.7
+  - 3.2
+  - 3.3
   - 3.4
   - 3.5
   - 3.6
