### 2018-02-01 8cba9e9
test on all Python versions Travis-CI supports
