### 2018-02-01 e40b305
add more functions
