+++ fortran/maptran.f90
@@ -10,8 +10,8 @@ module maptran
   end type
 
   real(wp), parameter :: pi = 4._wp * atan(1.0)
-  
-  public :: ecef2geodetic, geodetic2ecef, assert_isclose
+
+  public :: ecef2geodetic, geodetic2ecef, aer2enu, aer2ecef, assert_isclose
 
 contains
 
@@ -28,13 +28,12 @@ subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
 
   real(wp), intent(in) :: x,y,z
   type(wgs84Ellipsoid), intent(in), optional :: spheroid
-  logical, intent(in), optional :: deg
+  logical, optional,value :: deg
   real(wp), intent(out) :: lat, lon, alt
 
-  type(wgs84Ellipsoid) :: ell
-  logical dg
   real(wp) :: ea, eb, rad, rho, c, vnew, v
   integer :: i
+  type(wgs84Ellipsoid) :: ell
 
   if (present(spheroid)) then
      ell = spheroid
@@ -42,11 +41,7 @@ subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
      ell = wgs84Ellipsoid()
   endif
 
-  if (present(deg)) then
-      dg = deg
-  else
-      dg = .true.
-  endif
+  if (.not.present(deg)) deg = .true.
 
     ea = ell%SemimajorAxis
     eb = ell%SemiminorAxis
@@ -76,33 +71,112 @@ subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
     alt = ((rad - ea * cos(vnew)) * cos(lat)) + &
            ((z - eb * sin(vnew)) * sin(lat))
 
-    if (dg) lat = degrees(lat); lon = degrees(lon)
+    if (deg) lat = degrees(lat); lon = degrees(lon)
 
 end subroutine ecef2geodetic
 
 
 subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
-  real(wp), intent(in) :: lat,lon,alt
+  real(wp), value :: lat,lon
+  real(wp), intent(in) :: alt
   real(wp), intent(out) :: x,y,z
-  type(wgs84Ellipsoid) :: spheroid
-  logical, intent(in) :: deg
-  
-  real(wp) :: N, lat1,lon1
-  
-  if (deg) then
-    lat1 = radians(lat); lon1 = radians(lon)
+  type(wgs84Ellipsoid), intent(in), optional :: spheroid
+  logical, optional, value :: deg
+
+  real(wp) :: N
+  type(wgs84Ellipsoid) :: ell
+
+  if (present(spheroid)) then
+     ell = spheroid
   else
-    lat1 = lat; lon1 = lon
+     ell = wgs84Ellipsoid()
   endif
+
+  if (.not.present(deg)) deg=.true.
+
+  if (deg) lat = radians(lat); lon = radians(lon)
+
 ! Radius of curvature of the prime vertical section
-  N = radius_normal(lat1, spheroid)
+  N = radius_normal(lat, ell)
 ! Compute cartesian (geocentric) coordinates given  (curvilinear) geodetic coordinates.
-  
-  x = (N + alt) * cos(lat1) * cos(lon1)
-  y = (N + alt) * cos(lat1) * sin(lon1)
-  z = (N * (spheroid%SemiminorAxis / spheroid%SemimajorAxis)**2 + alt) * sin(lat1)
+
+  x = (N + alt) * cos(lat) * cos(lon)
+  y = (N + alt) * cos(lat) * sin(lon)
[…]
