### 2018-02-01 #560 @lheagy
:wave: Hi @leouieda would you be willing and interested to review this submission for JOSS?
