### 2018-02-01 0d5d4a7
add more Fortran functions--WIP
