+++ fortran/maptran.f90
@@ -11,7 +11,7 @@ module maptran
 
   real(wp), parameter :: pi = 4._wp * atan(1.0)
 
-  public :: ecef2geodetic, geodetic2ecef, aer2enu, aer2ecef, assert_isclose
+  public :: ecef2geodetic, geodetic2ecef, aer2enu, aer2ecef, ecef2aer, enu2ecef, ecef2enu, assert_isclose
 
 contains
 
@@ -108,7 +108,7 @@ end subroutine geodetic2ecef
 
 
 subroutine aer2ecef(az, el, slantRange, lat0, lon0, alt0, x,y,z, spheroid, deg)
-!er2ecef  convert azimuth, elevation, range to target from observer to ECEF coordinates
+! aer2ecef  convert azimuth, elevation, range to target from observer to ECEF coordinates
 !
 ! Inputs
 ! ------
@@ -145,6 +145,35 @@ subroutine aer2ecef(az, el, slantRange, lat0, lon0, alt0, x,y,z, spheroid, deg)
 end subroutine aer2ecef
 
 
+subroutine ecef2aer(x, y, z, lat0, lon0, alt0, az, el, slantRange, spheroid, deg)
+! ecef2aer  convert ECEF of target to azimuth, elevation, slant range from observer
+!
+! Inputs
+! ------
+! x,y,z: Earth Centered Earth Fixed (ECEF) coordinates of test point (meters)
+! lat0, lon0, alt0: ellipsoid geodetic coordinates of observer/reference (degrees, degrees, meters)
+! spheroid: referenceEllipsoid parameter struct
+! deg: .true.: degrees
+!
+! Outputs
+! -------
+! az, el, slantrange: look angles and distance to point under test (degrees, degrees, meters)
+! az: azimuth clockwise from local north
+! el: elevation angle above local horizon
+
+  real(wp), intent(in) :: x,y,z, lat0, lon0, alt0
+  type(wgs84Ellipsoid), intent(in), optional :: spheroid
+  logical, intent(in), optional :: deg
+  real(wp), intent(out) :: az,el, slantRange
+  
+  real(wp) :: east, north, up
+
+  call ecef2enu(x, y, z, lat0, lon0, alt0, east, north, up, spheroid, deg)
+  call enu2aer(east, north, up, az, el, slantRange, deg)
+  
+end subroutine ecef2aer
+
+
 subroutine aer2enu(az, el, slantRange, east, north, up, deg)
 !aer2enu  convert azimuth, elevation, range to ENU coordinates
 !
@@ -168,7 +197,7 @@ subroutine aer2enu(az, el, slantRange, east, north, up, deg)
 
   if(.not.present(deg)) deg=.true.
 
-  if (deg) az = degrees(az); el = degrees(el)
+  if (deg) az = radians(az); el = radians(el)
 
 ! Calculation of AER2ENU
    up = slantRange * sin(el)
@@ -179,8 +208,129 @@ subroutine aer2enu(az, el, slantRange, east, north, up, deg)
 end subroutine aer2enu
 
 
+subroutine enu2aer(east, north, up, az, elev, slantRange, deg)
+! enu2aer   convert ENU to azimuth, elevation, slant range
+!
+! Inputs
+! ------
+! e,n,u:  East, North, Up coordinates of test points (meters)
+! angleUnit: string for angular units. Default 'd': degrees
+!
+! outputs
+! -------
+! az, el, slantrange: look angles and distance to point under test (degrees, degrees, meters)
+! az: azimuth clockwise from local north
[…]
