+++ fortran/maptran.f90
@@ -3,15 +3,21 @@ module maptran
   implicit none
   private
 
-  type, public :: wgs84Ellipsoid
-     real(wp) :: SemimajorAxis = 6378137.  ! [m]
+  type,public :: referenceEllipsoid
+     real(wp) :: SemimajorAxis 
 !     real :: Flattening = 1 / 298.2572235630  ! flattening
-     real(wp) :: SemiminorAxis = 6378137. * (1 - 1 / 298.2572235630)
+     real(wp) :: SemiminorAxis 
   end type
 
   real(wp), parameter :: pi = 4._wp * atan(1.0)
+  
+  type(referenceEllipsoid), parameter, public :: wgs84Ellipsoid = &
+        referenceEllipsoid(SemimajorAxis=6378137., &
+                           SemiminorAxis=6378137. * (1 - 1 / 298.2572235630))
 
-  public :: ecef2geodetic, geodetic2ecef, aer2enu, aer2ecef, ecef2aer, enu2ecef, ecef2enu, assert_isclose
+  public :: ecef2geodetic, geodetic2ecef, aer2enu, enu2aer, aer2ecef, ecef2aer, &
+            enu2ecef, ecef2enu, aer2geodetic, geodetic2enu,assert_isclose, &
+            geodetic2aer,enu2geodetic
 
 contains
 
@@ -27,18 +33,18 @@ subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
 ! not involve division by cos(phi) or sin(phi)
 
   real(wp), intent(in) :: x,y,z
-  type(wgs84Ellipsoid), intent(in), optional :: spheroid
+  type(referenceEllipsoid), intent(in), optional :: spheroid
   logical, optional,value :: deg
   real(wp), intent(out) :: lat, lon, alt
 
   real(wp) :: ea, eb, rad, rho, c, vnew, v
   integer :: i
-  type(wgs84Ellipsoid) :: ell
+  type(referenceEllipsoid) :: ell
 
   if (present(spheroid)) then
      ell = spheroid
   else
-     ell = wgs84Ellipsoid()
+     ell = wgs84Ellipsoid
   endif
 
   if (.not.present(deg)) deg = .true.
@@ -80,16 +86,16 @@ subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
   real(wp), value :: lat,lon
   real(wp), intent(in) :: alt
   real(wp), intent(out) :: x,y,z
-  type(wgs84Ellipsoid), intent(in), optional :: spheroid
+  type(referenceEllipsoid), intent(in), optional :: spheroid
   logical, optional, value :: deg
 
   real(wp) :: N
-  type(wgs84Ellipsoid) :: ell
+  type(referenceEllipsoid) :: ell
 
   if (present(spheroid)) then
      ell = spheroid
   else
-     ell = wgs84Ellipsoid()
+     ell = wgs84Ellipsoid
   endif
 
   if (.not.present(deg)) deg=.true.
@@ -105,6 +111,129 @@ subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
   z = (N * (ell%SemiminorAxis / ell%SemimajorAxis)**2 + alt) * sin(lat)
 
 end subroutine geodetic2ecef
+
+
+subroutine aer2geodetic(az, el, slantRange, lat0, lon0, alt0, lat1, lon1, alt1, spheroid, deg)
+!aer2geodetic  convert azimuth, elevation, range of target from observer to geodetic coordiantes
+!
+! Inputs
+! ------
+! az, el, slantrange: look angles and distance to point under test (degrees, degrees, meters)
+! az: azimuth clockwise from local north
[…]
