### 2018-02-02 ba434b1
selftest pass fortran [ci skip]
