+++ fortran/Test.f90
@@ -6,7 +6,7 @@ use maptran
 implicit none
 
 real(wp), parameter :: lat = 42, lon= -82, alt = 200, &
-                       az = 33, el=70 ,srange = 1e3, &
+                       az = 33, el=70, rng= 1e3, &
                        x0 = 660.675e3, y0 = -4700.949e3, z0 = 4245.738e3, &
                        xl = 660.930e3, yl = -4701.424e3, zl = 4246.579e3, & ! aer2ecef
                        er = 186.277521, nr = 286.84222, ur = 939.69262, & ! aer2enu
@@ -25,12 +25,12 @@ call assert_isclose(x1,x0)
 call assert_isclose(y1,y0)
 call assert_isclose(z1,z0)
 
-call aer2enu(az, el, srange, e1, n1, u1)
+call aer2enu(az, el, rng, e1, n1, u1)
 call assert_isclose(e1, er)
 call assert_isclose(n1, nr)
 call assert_isclose(u1, ur)
 
-call aer2ecef(az,el,srange,lat,lon,alt,x2,y2,z2)
+call aer2ecef(az,el,rng,lat,lon,alt,x2,y2,z2)
 call assert_isclose(x2, xl)
 call assert_isclose(y2, yl)
 call assert_isclose(z2, zl)
@@ -41,12 +41,16 @@ call assert_isclose(lon2,lon)
 call assert_isclose(alt2,alt)
 
 call enu2aer(e1,n1,u1, az2, el2, rng2)
+call assert_isclose(az2,az)
+call assert_isclose(el2,el)
+call assert_isclose(rng2,rng)
 
 call ecef2aer(x2,y2,z2, lat,lon,alt, az3, el3, rng3)
-call assert_isclose(el3, el)
 call assert_isclose(az3, az)
+call assert_isclose(el3, el)
+call assert_isclose(rng3,rng)
 
-call aer2geodetic(az,el,srange,lat,lon,alt, lat3,lon3,alt3)
+call aer2geodetic(az,el,rng,lat,lon,alt, lat3,lon3,alt3)
 call assert_isclose(lat3, lat1)
 
 call geodetic2enu(lat3, lon3, alt3, lat, lon, alt, e2, n2, u2)
