+++ fortran/maptran.f90
@@ -3,7 +3,7 @@ module maptran
   implicit none
   private
 
-  type,public :: referenceEllipsoid
+  type,public :: Ellipsoid
      real(wp) :: SemimajorAxis 
      real(wp) :: Flattening 
      real(wp) :: SemiminorAxis 
@@ -11,10 +11,10 @@ module maptran
 
   real(wp), parameter :: pi = 4._wp * atan(1.0_wp)
   
-  type(referenceEllipsoid), parameter, public :: wgs84Ellipsoid = &
-        referenceEllipsoid(SemimajorAxis=6378137._wp, &
-                           SemiminorAxis=6378137._wp * (1._wp - 1._wp / 298.2572235630_wp), &
-                           Flattening = 1. / 298.2572235630_wp)
+  type(Ellipsoid), parameter, public :: wgs84Ellipsoid = &
+        Ellipsoid(SemimajorAxis=6378137._wp, &
+                  SemiminorAxis=6378137._wp * (1._wp - 1._wp / 298.2572235630_wp), &
+                  Flattening = 1. / 298.2572235630_wp)
 
   public :: ecef2geodetic, geodetic2ecef, aer2enu, enu2aer, aer2ecef, ecef2aer, &
             enu2ecef, ecef2enu, aer2geodetic, geodetic2enu,assert_isclose, &
@@ -34,13 +34,13 @@ subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
 ! not involve division by cos(phi) or sin(phi)
 
   real(wp), intent(in) :: x,y,z
-  type(referenceEllipsoid), intent(in), optional :: spheroid
+  type(Ellipsoid), intent(in), optional :: spheroid
   logical, optional,value :: deg
   real(wp), intent(out) :: lat, lon, alt
 
   real(wp) :: ea, eb, rad, rho, c, vnew, v
   integer :: i
-  type(referenceEllipsoid) :: ell
+  type(Ellipsoid) :: ell
 
   if (present(spheroid)) then
      ell = spheroid
@@ -89,7 +89,7 @@ subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
 ! Inputs
 ! ------
 ! lat,lon, alt:  ellipsoid geodetic coordinates of point(s) (degrees, degrees, meters)
-! spheroid: referenceEllipsoid parameter struct
+! spheroid: Ellipsoid parameter struct
 ! deg: .true. degrees
 !
 ! outputs
@@ -99,11 +99,11 @@ subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
   real(wp), value :: lat,lon
   real(wp), intent(in) :: alt
   real(wp), intent(out) :: x,y,z
-  type(referenceEllipsoid), intent(in), optional :: spheroid
+  type(Ellipsoid), intent(in), optional :: spheroid
   logical, optional, value :: deg
 
   real(wp) :: N
-  type(referenceEllipsoid) :: ell
+  type(Ellipsoid) :: ell
 
   if (present(spheroid)) then
      ell = spheroid
@@ -135,7 +135,7 @@ subroutine aer2geodetic(az, el, slantRange, lat0, lon0, alt0, lat1, lon1, alt1, 
 ! az: azimuth clockwise from local north
 ! el: elevation angle above local horizon
 ! lat0, lon0, alt0: ellipsoid geodetic coordinates of observer/reference (degrees, degrees, meters)
-! spheroid: referenceEllipsoid parameter struct
+! spheroid: Ellipsoid parameter struct
 ! deg: .true.: degrees
 !
 ! Outputs
[…]
