+++ fortran/maptran.f90
@@ -1,5 +1,5 @@
 module maptran
-  use iso_fortran_env, only: wp=>real64
+  use, intrinsic:: iso_fortran_env, only: wp=>real64
   implicit none
   private
 
@@ -16,13 +16,14 @@ module maptran
                   SemiminorAxis=6378137._wp * (1._wp - 1._wp / 298.2572235630_wp), &
                   Flattening = 1. / 298.2572235630_wp)
 
-  public :: ecef2geodetic, geodetic2ecef, aer2enu, enu2aer, aer2ecef, ecef2aer, &
+  public :: wp,ecef2geodetic, geodetic2ecef, aer2enu, enu2aer, aer2ecef, ecef2aer, &
             enu2ecef, ecef2enu, aer2geodetic, geodetic2enu,assert_isclose, &
-            geodetic2aer,enu2geodetic
+            geodetic2aer,enu2geodetic,degrees,radians
 
 contains
 
-subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
+
+elemental subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
 
 ! convert ECEF (meters) to geodetic coordintes
 
@@ -83,7 +84,7 @@ subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
 end subroutine ecef2geodetic
 
 
-subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
+elemental subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
 !geodetic2ecef   convert from geodetic to ECEF coordiantes
 !
 ! Inputs
@@ -103,8 +104,8 @@ subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
   logical, optional, value :: deg
 
   real(wp) :: N
-  type(Ellipsoid) :: ell
-
+  type(Ellipsoid) :: ell 
+  
   if (present(spheroid)) then
      ell = spheroid
   else
@@ -126,7 +127,7 @@ subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
 end subroutine geodetic2ecef
 
 
-subroutine aer2geodetic(az, el, slantRange, lat0, lon0, alt0, lat1, lon1, alt1, spheroid, deg)
+elemental subroutine aer2geodetic(az, el, slantRange, lat0, lon0, alt0, lat1, lon1, alt1, spheroid, deg)
 !aer2geodetic  convert azimuth, elevation, range of target from observer to geodetic coordiantes
 !
 ! Inputs
@@ -155,7 +156,7 @@ subroutine aer2geodetic(az, el, slantRange, lat0, lon0, alt0, lat1, lon1, alt1, 
 end subroutine aer2geodetic
 
 
-subroutine geodetic2aer(lat, lon, alt, lat0, lon0, alt0, az, el, slantRange, spheroid, deg)
+elemental subroutine geodetic2aer(lat, lon, alt, lat0, lon0, alt0, az, el, slantRange, spheroid, deg)
 !geodetic2aer   from an observer's perspective, convert target coordinates to azimuth, elevation, slant range.
 !
 ! Inputs
@@ -186,7 +187,7 @@ end subroutine geodetic2aer
 
 
 
-subroutine geodetic2enu(lat, lon, alt, lat0, lon0, alt0, east, north, up, spheroid, deg)
+elemental subroutine geodetic2enu(lat, lon, alt, lat0, lon0, alt0, east, north, up, spheroid, deg)
 ! geodetic2enu    convert from geodetic to ENU coordinates
 !
 ! Inputs
[…]
