+++ fortran/Test.f90
@@ -1,9 +1,10 @@
 program Test
 
-use iso_fortran_env, only: wp=>real64
 use maptran
 
 implicit none
+
+!type(Ellipsoid), parameter :: spheroid = wgs84Ellipsoid
 
 real(wp), parameter :: lat = 42, lon= -82, alt = 200, &
                        az = 33, el=70, rng= 1e3, &
@@ -11,78 +12,83 @@ real(wp), parameter :: lat = 42, lon= -82, alt = 200, &
                        xl = 660.930e3, yl = -4701.424e3, zl = 4246.579e3, & ! aer2ecef
                        er = 186.277521, nr = 286.84222, ur = 939.69262, & ! aer2enu
                        lat1 = 42.0026, lon1 = -81.9978, alt1 = 1.1397e3 ! aer2geodetic
+               
+integer,parameter :: N = 3               
+real(wp), dimension(N), parameter :: alat = [42,52,62], &
+                                     deg0 = [15,30,45], &
+                                     aaz = [33,43,53]
 
 real(wp) :: lat2, lon2, alt2,lat3,lon3,alt3,lat4,lon4,alt4,&
             x1,y1,z1,x2,y2,z2,x3,y3,z3,&
             az2,el2,rng2,az3,el3,rng3,az4,el4,rng4,&
             e1,n1,u1,e2,n2,u2,e3,n3,u3
+  
 
-type(Ellipsoid), parameter :: spheroid = wgs84Ellipsoid
+real(wp), dimension(N) :: ax1, ay1, aaaz1, ax2, ay2, aaaz2, ax3,ay3,aaaz3, &
+                          ae1, an1, au1, ae2,an2,au2, &
+                          alat2,alon2,aalt2, alat3,alon3,aalt3, alat4,alon4,aalt4, &
+                          aaz2, ael2, arng2, aaz3,ael3,arng3, aaz4,ael4,arng4
 
-print*,'Default WGS84 Ellipsoid:',spheroid
+real(wp), dimension(N) :: rad,deg
 
+
+!print*,'Default WGS84 Ellipsoid:',spheroid
+
+! --------- scalar
 
 call geodetic2ecef(lat,lon,alt,x1,y1,z1)
-call assert_isclose(x1,x0)
-call assert_isclose(y1,y0)
-call assert_isclose(z1,z0)
+call assert_isclose([x1,y1,z1],[x0,y0,z0])
 
 call aer2enu(az, el, rng, e1, n1, u1)
-call assert_isclose(e1, er)
-call assert_isclose(n1, nr)
-call assert_isclose(u1, ur)
+call assert_isclose([e1,n1,u1], [er,nr,ur])
 
 call aer2ecef(az,el,rng,lat,lon,alt,x2,y2,z2)
-call assert_isclose(x2, xl)
-call assert_isclose(y2, yl)
-call assert_isclose(z2, zl)
+call assert_isclose([x2,y2,z2],[xl,yl,zl])
 
 call ecef2geodetic(x1,y1,z1,lat2,lon2,alt2)
-call assert_isclose(lat2,lat)
-call assert_isclose(lon2,lon)
-call assert_isclose(alt2,alt)
+call assert_isclose([lat2,lon2,alt2],[lat,lon,alt])
 
 call enu2aer(e1,n1,u1, az2, el2, rng2)
-call assert_isclose(az2,az)
-call assert_isclose(el2,el)
-call assert_isclose(rng2,rng)
[…]
