+++ CMakeLists.txt
@@ -1,7 +1,8 @@
 cmake_minimum_required(VERSION 2.8.12)
 project(MapTran Fortran)
 
-add_compile_options(-ffpe-trap=invalid,zero,overflow -g -fbacktrace -Wall -Wextra -Warray-bounds)
+add_compile_options(-std=f2008 -ffpe-trap=invalid,zero,overflow -g -fbacktrace 
+                    -Wall -Wextra -Warray-bounds -Wfatal-errors)
 
 add_library(maptran fortran/maptran.f90)
 
