### 2018-02-02 bb9fc20
fortran selftest
