### 2018-02-11 #580 @whedon
[REVIEW]: PyMap3D: 3-D coordinate conversions for terrestrial and geospace environments
**Submitting author:** @scivision (<a href="http://orcid.org/0000-0002-1637-6526">Michael Hirsch</a>)
**Repository:** <a href="https://github.com/scivision/pymap3d" target ="_blank">https://github.com/scivision/pymap3d</a>
**Version:** v1.6.0
