### 2018-02-11 #580 @lheagy
:wave: Hi @hugoledoux and @leouieda, here is the issue for the review. I have put 2 checklists in the top of the issue, one for each of you. Please let me know if you have any questions!
