### 2018-02-15 4af7c1c
robust assert template
