+++ fortran/vallado.f90
@@ -6,60 +6,83 @@ private
 integer, parameter :: wp=dp
 real(wp), parameter :: pi = 4._wp * atan(1.0_wp)
 
-public:: gstime, juliandayall, lstime
+public:: toGST, toJulian, toLST, radec2azel, azel2radec
 contains
 
+elemental subroutine azel2radec(az,el,lat,lon,jd, ra,decl)
 
-elemental SUBROUTINE RADEC_AZEL  ( RtAsc,Decl,LST,Latgd, Direction, Az,El )
+  real(wp), value :: az,el,lat,lon
+  real(wp), intent(in) :: jd ! Julian Date
+  real(wp), intent(out) :: ra, decl
+  real(wp) :: lst, lha, sinv, cosv
+  
+  az = radians(az)
+  el = radians(el)
+  lat = radians(lat)
+  lon = radians(lon)
 
-  REAL(wp), value :: RtAsc,Decl,LST,Latgd
+  Decl = ASIN(SIN(El)*SIN(lat) + &
+              COS(el)*COS(lat)*COS(Az) )
+
+  Sinv = -(SIN(az)*COS(el)*COS(lat)) / &
+          (COS(lat)*COS(Decl))
+          
+  Cosv = (SIN(el) - SIN(lat)*SIN(decl)) / &
+         (COS(lat)*COS(Decl))
+         
+  LHA  = ATAN2(Sinv,Cosv)
+  lst = toLST(Lon, JD)
+  
+  ra = modulo(degrees(LST - LHA), 360._wp)
+  decl = degrees(decl)
+
+end subroutine azel2radec
+
+
+elemental SUBROUTINE radec2azel(ra,Decl,lat,lon,jd, Az,El)
+
+  REAL(wp), value :: ra,Decl,lat, lon
+  real(wp), intent(in) :: jd
   real(wp), intent(out) :: Az,El
-  CHARACTER(4), intent(in), optional :: Direction
   REAL(wp) :: Sinv, Cosv, LHA
+  
+  lat = radians(lat)
+  lon = radians(lon)
+  ra = radians(ra)
+  decl = radians(decl)
 
-        IF (present(direction) .and. Direction .eq. 'FROM' ) THEN
-            Decl = ASIN( SIN(El)*SIN(Latgd) + &
-     &                 COS(el)*COS(Latgd)*COS(Az) )
+  LHA = toLST(Lon, JD) - ra
 
-            Sinv = -(SIN(az)*COS(el)*COS(Latgd)) / &
-     &              (COS(Latgd)*COS(Decl))
-            Cosv = (SIN(el) - SIN(Latgd)*SIN(decl)) / &
-     &              (COS(Latgd)*COS(Decl))
-            LHA  = ATAN2( Sinv,Cosv ) 
-            RtAsc= LST - LHA 
-          ELSE
-            LHA = LST - RtAsc
+  El = ASIN( SIN(Decl)*SIN(lat) + &
+              COS(Decl)*COS(lat)*COS(LHA) )
 
-            El  = ASIN( SIN(Decl)*SIN(Latgd) + &
-     &            COS(Decl)*COS(Latgd)*COS(LHA) )
-
-            Sinv= -SIN(LHA)*COS(Decl)*COS(Latgd) / &
-     &                (COS(el)*COS(Latgd))
-            Cosv= ( SIN(Decl)-SIN(el)*SIN(Latgd) ) / &
-     &             (COS(el)*COS(Latgd))
-            Az  = ATAN2( Sinv,Cosv ) 
-          ENDIF 
+  Sinv = -SIN(LHA)*COS(Decl)*COS(lat) / &
+        (COS(el)*COS(lat))
+        
[…]
