### 2018-02-20 29a5b3e
correct variable/function names
