### 2018-02-22 #580 @leouieda
Hi @lheagy, I can get started on this tomorrow. I had a look at the guidelines and review questions.

It seems that there has been a new release (1.6.0) since the paper was submitted. Should @scivision update the paper/issue content to match?
