### 2018-02-22 #580 @lheagy
Thanks @leouieda, yes, that is a good catch. @arfon: how do we update the version on this issue? manually? or should this be done with whedon?
