### 2018-02-22 #580 @arfon
> @arfon: how do we update the version on this issue? manually? or should this be done with whedon?

You can simply edit the comment by @whedon at the top of the issue.
