### 2018-02-23 #580 @scivision
Thanks, I updated [paper.md](https://github.com/scivision/pymap3d/commit/c07feead980407e622da9086ce11df2d4174eb96) to mention the Fortran code. The Matlab and Fortran code are completely independent standalone codes for those langauges.
