### 2018-02-27 #580 @hugoledoux
@scivision

I am a GIS person used to work with proj4 (pyproj), and the Python package works as advertised and is surely useful to people in that community. For a GIS person, proj4 is "standard" and everyone uses it.

I only tested the Python version, since I don't have the 1000$ to buy Matlab...

## README
I find the README a bit confusing sometimes to be honest.

## install
You state in INSTALL that there are 3 packages, but you give two installations...

## prereqs
I'd write "prerequisites", but okay fair enough.

What is not clear to me: if I want to use numpy for instance, do I have to do something? It's on my system, is it used automatically? Same thing for AstroPy.

Also a link to AstroPy could help the user.

## version
The version is 1.6.0, I changed the from 1.5.0. I assume this is not a problem.

## pyproj
I believe there are conversions that you perform that are impossible with proj4, no? State a few... For me the argument of not pure python is not very strong, it's trivial to install proj4 on most platform.

## installation
I think it could be written that everything is on PyPI and just doing `pip install pymap3d` works (that's how I installed it under Python 2.7)

## functions
I am not familiar with the terms such as 'aer', 'ned', etc. Are these explained somewhere? That is needed in my opinion, now the docs doesn't really explain them either.

## Software paper
- geospace coordinate systems --> *geographic* coordinate systems is a better term in my opinion

## usage
Not totally clear why there are numbers there `az,el,range = pm.geodetic2aer(lat, lon, alt, 42, -82, 0)`, I mean the 42 and -82. Perhaps just explain?

## community guidelines
I don't see any file there.
