### 2018-02-28 scivision/pymap3d#7 @leouieda
Tag for v1.6.0 has an extra .
Part of openjournals/joss-reviews#580

The tag for the latest version is `v.1.6.0` instead of `v1.6.0`.
