### 2018-02-28 scivision/pymap3d#8 @leouieda
Install instructions
Part of openjournals/joss-reviews#580

It would be good to have more detailed install instructions. Things I found missing:
* How to download the source (command to clone the repository or link to download a zip archive from Github)
* How to install the Matlab version

Regarding the Python instructions, `pip install -e` is used for development but not recommended for deploying software because the user can't delete the source directory after install. I would recommend one of the following:
1. Make a release and upload to PyPI. Then the instructions should be: `pip install pymap3d`
2. Download the zip/clone the repo and install using pip: `cd pymap3d; pip install .` (no `-e`)
3. Install directly from master using pip: `pip install https://github.com/scivision/pymap3d/archive/master.zip`

1 is the best option but takes a bit more work to setup. With 3, you can also include instructions for installing specific versions using the same syntax but replacing `master` with the tag name (see #7). The problem with 2 and 3 is that they don't play well with dependency managers. So it's not easy for another package to depend on pymap3d.
