### 2018-03-05 162841f
clarify readme, use pyproj for a few test comparison (optional), @hugoledoux comments
