+++ README.rst
@@ -18,7 +18,7 @@
 Python / Matlab / Fortran 3-D coordinate conversions
 ====================================================
 
-3-D coordinate conversions, like API of popular $1000 Matlab Mapping Toolbox routines for:
+3-D geographic coordinate conversions, with API similar to popular $1000 Matlab Mapping Toolbox routines for:
 
 * Python
 * Matlab, GNU Octave
@@ -30,39 +30,51 @@ Includes some relevant `Vallado's algorithms <http://www.smad.com/vallado/fortra
 
 For those not having:
 
-* AstroPy: lower accuracy fallback functions are included for some functions.
-* Numpy: without numpy, scalar inputs are handled with pure Python builtins.
+* `AstroPy <http://www.astropy.org/>`_: lower accuracy fallback functions are automatically used.
+* Numpy: without Numpy, scalar inputs are handled with pure Python builtins.
 
-Why not `PyProj <https://github.com/jswhit/pyproj>`_? 
-PyMap3D does not require anything beyond pure Python.
-PyProj is very powerful and comprehensive, but requires a learning curve to use, 
-particularly for Python users accustomed to Matlab.
+Why not `PyProj <https://github.com/jswhit/pyproj>`_?
+
+* PyMap3D does not require anything beyond pure Python.
+* PyMap3D API is virtually identical to Matlab Mapping Toolbox, while PyProj's interface is quite distinct
+* PyMap3D intrinsically handles local coordinate systems such as ENU, while for PyProj ENU requires some `additional effort <https://github.com/jswhit/pyproj/issues/105>`_.
+* PyProj is oriented towards points on the planet surface, while PyMap3D handles points on or above the planet surface equally well, particularly import for airborne vehicles and remote sensing.
 
 .. contents::
 
 
-Prereqs
-=======
+Prerequisites
+=============
 
 * Python PyMap3D:  any of Python 2.6, 2.7, 3.3, 3.4, 3.5, 3.6, 3.7, ...
-  * optional: Numpy, AstroPy  (for full functionality)
+  * Numpy (optional): if not present, function from ``math`` are automatically used, and if Numpy is present, it is automatically used.
+  * `AstroPy <http://www.astropy.org/>`_  (optional): If not present, ECI coordinate conversions are not available.
 * Matlab / GNU Octave: under ``matlab/``
 * Fortran MapTran: under ``fortran/``:  any Fortran compiler (tested with ``gfortran``)
 
 Install
 =======
-This repo is 3 separate packages, you can use them independently, they don't rely on each other.
+The three separate packages are independent, they don't rely on each other.
 
 * Python PyMap3D::
-    
-    python -m pip install -e .
-    
+
+      pip install pymap3d
+
+  or for the latest development code::
+
+      git clone https://github.com/scivision/pymap3d
+
+      pip install -e .
+
 * Fortran MapTran::
 
     cd bin
     cmake ..
     make
- 
+
+* Matlab/Octave::
+
+    addpath(pymap3d/matlab)
 
 
 Usage
@@ -76,9 +88,9 @@ Python
    import pymap3d as pm
 
    x,y,z = pm.geodetic2ecef(lat,lon,alt)
-   
-   az,el,range = pm.geodetic2aer(lat, lon, alt, 42, -82, 0)
-   
+
+   az,el,range = pm.geodetic2aer(lat, lon, alt, observer_lat, observer_lon, 0)
+
 Matlab / GNU Octave
 -------------------
[…]
