+++ .travis.yml
@@ -34,8 +34,11 @@ addons:
 install: pip -q install -e .[tests]
 
 script: 
-  - coverage run tests/Test.py -v
+  - pytest
+
   - if [[ $TRAVIS_PYTHON_VERSION == 3.7* ]]; then
+      coverage run tests/test_all.py -v;
+      
       cd tests;
       octave-cli -q Test.m;
       cd ..;
