### 2018-03-05 scivision/pymap3d#8 @scivision
I added this information, it is already in PyPi, which is used by several programs to auto-install pymap3d when the other programs are installed.
