### 2018-03-05 #580 @scivision
These issues have been addressed in the latest git commits, thanks
