### 2018-03-05 5f78aa2
keep Travis-CI tests to actually used Python versions
