### 2018-03-06 scivision/pymap3d#8 @lheagy
@leouieda: would you mind taking a look at the updates here?
