### 2018-03-06 51e59db
Update README.rst
markdown list wasn't parsed
