### 2018-03-06 #580 @hugoledoux
Everything was done, except I don't see anything for this point "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

If it's done then all good with me.
