### 2018-03-06 b132b14
Create CONTRIBUTING.md
