### 2018-03-06 #580 @scivision
Thank you this has been added in
https://github.com/scivision/pymap3d/blob/master/docs/CONTRIBUTING.md
