### 2018-03-06 74d8248
Fortran radians bugfix, allow Matlab and Fortran assert to have err_msg, simplify Python test arguments
Squashed commit of the following:

commit 26c10026d8a5532e033cd5c77ac6f7f60b5398a2
Author: Michael Hirsch, Ph.D <scivision@users.noreply.github.com>
Date:   Tue Mar 6 12:23:14 2018 -0500

    add err_msg to assert

commit 125e786740471528c3e468f64335a26a1f602164
Author: Michael Hirsch, Ph.D <scivision@users.noreply.github.com>
Date:   Tue Mar 6 12:17:03 2018 -0500

    add more python radians tests

commit fe89b1ee8eea9d138dd126b7ef563161705edecd
Author: Michael Hirsch, Ph.D <scivision@users
[…]
