+++ fortran/Test.f90
@@ -35,17 +35,15 @@ real(wp), dimension(N) :: ax1, ay1, aaaz1, ax2, ay2, aaaz2, ax3,ay3,aaaz3, &
                           aaz2, ael2, arng2, aaz3,ael3,arng3, aaz4,ael4,arng4
                          
 
-real(wp), dimension(N) :: rad,deg
-
 
 !print*,'Default WGS84 Ellipsoid:',spheroid
 
-! --------- scalar
+! --------- scalar degrees
 
-call geodetic2ecef(lat,lon,alt,x1,y1,z1)
+call geodetic2ecef(lat,lon,alt, x1,y1,z1)
 call assert_isclose([x1,y1,z1],[x0,y0,z0])
 
-call aer2enu(az, el, rng, e1, n1, u1)
+call aer2enu(az,el,rng, e1,n1,u1)
 call assert_isclose([e1,n1,u1], [er,nr,ur])
 
 call aer2ecef(az,el,rng,lat,lon,alt,x2,y2,z2)
@@ -78,11 +76,48 @@ call assert_isclose([lat4,lon4,alt4],[lat3,lon3,alt3])
 call ecef2enu(x3,y3,z3,lat,lon,alt, e3,n3,u3)
 call assert_isclose([e3,n3,u3],[e1,n1,u1])
 
+call assert_isclose(degrees(radians(deg0)), deg0)
+
+! ------ scalar radians
+
+call geodetic2ecef(radians(lat),radians(lon),alt, x1,y1,z1, deg=.false.)
+call assert_isclose([x1,y1,z1],[x0,y0,z0])
+
+call aer2enu(radians(az),radians(el),rng, e1,n1,u1, deg=.false.)
+call assert_isclose([e1,n1,u1], [er,nr,ur])
+
+call aer2ecef(radians(az),radians(el),rng, radians(lat),radians(lon),alt, x2,y2,z2, deg=.false.)
+call assert_isclose([x2,y2,z2],[xl,yl,zl])
+
+call ecef2geodetic(x1,y1,z1,lat2,lon2,alt2, deg=.false.)
+call assert_isclose([degrees(lat2),degrees(lon2),alt2],[lat,lon,alt])
+
+call enu2aer(e1,n1,u1, az2, el2, rng2, deg=.false.)
+call assert_isclose([degrees(az2),degrees(el2),rng2],[az,el,rng], err_msg='enu2aer: rad')
+
+call ecef2aer(x2,y2,z2, radians(lat),radians(lon),alt, az3,el3,rng3, deg=.false.)
+call assert_isclose([degrees(az3),degrees(el3),rng3],[az,el,rng])
+
+call aer2geodetic(radians(az),radians(el),rng, radians(lat),radians(lon),alt, lat3,lon3,alt3, deg=.false.)
+call assert_isclose([degrees(lat3),degrees(lon3),alt3],[lat1,lon1,alt1], err_msg='aer2geodetic: rad')
+
+call geodetic2enu(lat3,lon3,alt3, radians(lat),radians(lon),alt, e2,n2,u2, deg=.false.)
+call assert_isclose([e2,n2,u2],[e1,n1,u1], err_msg='geodetic2enu: rad')
+
+call geodetic2aer(lat3,lon3,alt3, radians(lat),radians(lon),alt, az4,el4,rng4, deg=.false.)
+call assert_isclose([degrees(az4),degrees(el4),rng4],[az,el,rng], err_msg='geodetic2aer: rad')
+
+call enu2ecef(e1,n1,u1, radians(lat),radians(lon),alt, x3,y3,z3, deg=.false.)
[…]
