+++ fortran/maptran.f90
@@ -42,48 +42,47 @@ elemental subroutine ecef2geodetic(x, y, z, lat, lon, alt, spheroid, deg)
   integer :: i
   type(Ellipsoid) :: ell
 
-  if (present(spheroid)) then
-     ell = spheroid
-  else
-     ell = wgs84Ellipsoid
-  endif
+  ell = merge(spheroid, wgs84Ellipsoid, present(spheroid))
 
   if (.not.present(deg)) deg = .true.
 
-    ea = ell%SemimajorAxis
-    eb = ell%SemiminorAxis
-    rad = hypot(x, y)
+  ea = ell%SemimajorAxis
+  eb = ell%SemiminorAxis
+  rad = hypot(x, y)
 ! Constant required for Latitude equation
-    rho = atan2(eb * z, ea * rad)
+  rho = atan2(eb * z, ea * rad)
 ! Constant required for latitude equation
-    c = (ea**2 - eb**2) / hypot(ea * rad, eb * z)
+  c = (ea**2 - eb**2) / hypot(ea * rad, eb * z)
 ! Starter for the Newtons Iteration Method
-    vnew = atan2(ea * z, eb * rad)
+  vnew = atan2(ea * z, eb * rad)
 ! Initializing the parametric latitude
-    v = 0._wp
-    do i = 1,5
-      v = vnew
-     ! Newtons Method for computing iterations
-      vnew = v - ((2 * sin(v - rho) - c * sin(2 * v)) / &
-                (2 * (cos(v - rho) - c * cos(2 * v))))
+  v = 0._wp
+  do i = 1,5
+    v = vnew
+   ! Newtons Method for computing iterations
+    vnew = v - ((2 * sin(v - rho) - c * sin(2 * v)) / &
+              (2 * (cos(v - rho) - c * cos(2 * v))))
 
-      if (isclose(v,vnew)) exit
-    enddo
+    if (isclose(v,vnew)) exit
+  enddo
 
 ! Computing latitude from the root of the latitude equation
-    lat = atan2(ea * tan(vnew), eb)
+  lat = atan2(ea * tan(vnew), eb)
 ! by inspection
-    lon = atan2(y, x)
+  lon = atan2(y, x)
 
-    alt = ((rad - ea * cos(vnew)) * cos(lat)) + &
-           ((z - eb * sin(vnew)) * sin(lat))
+  alt = ((rad - ea * cos(vnew)) * cos(lat)) + &
+         ((z - eb * sin(vnew)) * sin(lat))
 
-    if (deg) lat = degrees(lat); lon = degrees(lon)
+  if (deg) then
+    lat = degrees(lat)
+    lon = degrees(lon)
+  endif
 
 end subroutine ecef2geodetic
 
 
-elemental subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
+elemental subroutine geodetic2ecef(lat,lon,alt, x,y,z, spheroid, deg)
 ! geodetic2ecef   convert from geodetic to ECEF coordiantes
 !
 ! Inputs
@@ -105,15 +104,14 @@ elemental subroutine geodetic2ecef(lat,lon,alt,x,y,z,spheroid,deg)
   real(wp) :: N
   type(Ellipsoid) :: ell 
   
-  if (present(spheroid)) then
-     ell = spheroid
-  else
-     ell = wgs84Ellipsoid
-  endif
+  ell = merge(spheroid, wgs84Ellipsoid, present(spheroid)) 
 
   if (.not.present(deg)) deg=.true.
 
-  if (deg) lat = radians(lat); lon = radians(lon)
+  if (deg) then
+    lat = radians(lat)
[…]
