### 2018-03-06 698d3ff
cleanup Vallado API, make test parameters same as Python
