### 2018-03-06 d72bbbd
note that Python>=3.5 needed for optinal convenience argument unpacking
