+++ .travis.yml
@@ -2,13 +2,9 @@ language: python
 fast_finish: true
 
 python:
-  - 2.6
-  - 2.7
-  - 3.4
   - 3.5
   - 3.6
   
-
 os:
   - linux
   
