### 2018-03-07 c09b5f2
Add AppVeyor Windows and Ubuntu 16.04 self-test of Fortran and Python [ci skip]
Squashed commit of the following:

commit 297cc24d07165837e24b79df68a62688001becf2
Author: Michael Hirsch, Ph.D <scivision@users.noreply.github.com>
Date:   Wed Mar 7 15:40:01 2018 -0500

    test fortran and python

commit 4e26359db3dbe08fdb01b2e44027153080e7fe2b
Author: Michael Hirsch, Ph.D <scivision@users.noreply.github.com>
Date:   Wed Mar 7 15:37:43 2018 -0500

    init appveyor
