### 2018-03-08 scivision/pymap3d#10 @leouieda
Example usage with real values
Part of openjournals/joss-reviews#580

It's great that there are examples in the README. They would even more helpful if they included actual values and the expected outcomes so that we can check if we have the correct results.

This is not a requirement for the review, just a suggestion.
