### 2018-03-08 scivision/pymap3d#11 @leouieda
Parameter descriptions in the API docs
Part of openjournals/joss-reviews#580

The API docs are complete and probably satisfy most advanced users who are familiar with all the terms and units involved. They can be made more accessible if the function docstrings also include a description of each input and output parameter with units and/or data types and possible accepted values.

For example, it's not clear from the [docstring of ecef2geodetic](https://www.scivision.co/pymap3d/html/index.html#pymap3d.ecef2geodetic) how to specify an ellipsoid (I'm assuming that's what `ell` is) or what happens if `deg=False`.

This is not a requirement for the review, just a suggestion.
