### 2018-03-08 scivision/pymap3d#12 @leouieda
Instructions for running the automated tests
Part of openjournals/joss-reviews#580

I see that there are tests and building the fortran library creates a `testmaptran` program that seems to be run fortran tests. But I couldn't find instructions for doing this anywhere in the README, docs or CONTRIBUTING.md.

I found the commands in the `.travis.yml` file but this is not accessible to users unfamiliar with continuous integration. Running the Python tests is not trivial because the package needs to be installed with `[tests]` apparently and this adds a bunch of extra dependencies.

Even so, I was able to run all the tests on my machine without any problems. It would be nice to have these instructions in the README or CONTRIBUTING files.
