### 2018-03-08 scivision/pymap3d#13 @leouieda
Make CONTRIBUTING.md more visible
Part of openjournals/joss-reviews#580

It's great that the project has a contributing guide. Well done!

Currently, the file a bit difficult to find and not very visible. It took me a while to find it and I only looked in the docs folder because I saw the commit message that added it. I would suggest moving it to the repository root and/or including a reference to it in the README so that it can be easily found.

This is not a requirement for the review, just a suggestion.
