### 2018-03-09 #580 @lheagy
Thanks @hugoledoux and @leouieda for the review! @leouieda: there is still one outstanding check-box in the Documentation: the community guidelines. Did this just get missed?
