### 2018-03-12 scivision/pymap3d#10 @scivision
I added to the Installation section the commands for each language's self-test routine. It tests every function with degrees and radians, round-trip conversion and some independent cross-checks with PROJ.4
