### 2018-03-12 434078c
add self-test commands [ci skip]
