### 2018-03-12 9a4cdd4
docstrings [ci skip]
