+++ pymap3d/__init__.py
@@ -38,7 +38,7 @@ from .timeconv import str2dt
 
 
 class EarthEllipsoid:
-
+    """generate reference ellipsoid"""
     def __init__(self,model='wgs84'):
         if model == 'wgs84':
             """https://en.wikipedia.org/wiki/World_Geodetic_System#WGS84"""
@@ -55,8 +55,20 @@ class EarthEllipsoid:
 #%% to AER (azimuth, elevation, range)
 def ecef2aer(x, y, z, lat0, lon0, h0, ell=None, deg=True):
     """
-    Convert ECEF of target to azimuth, elevation, range (meters) from observer at lat0,lon0,h0 (meters)
-    output: azimuth (deg), elevation (deg), slant range (m) for Obs->Point
+    Observer => Point
+
+    input:
+    -----
+    x,y,z  [meters] target ECEF location                             [0,Infinity)
+    lat0, lon0 (degrees/radians)  Observer coordinates on ellipsoid  [-90,90],[-180,180]
+    h0     [meters]                observer altitude                 [0,Infinity)
+    ell    reference ellipsoid
+    deg    degrees input/output  (False: radians in/out)
+
+    output: AER
+    ------
+    azimuth, elevation (degrees/radians)                             [0,360),[0,90]
+    slant range [meters]                                             [0,Infinity)
     """
     xEast, yNorth, zUp = ecef2enu(x, y, z, lat0, lon0, h0, ell, deg=deg)
 
@@ -65,8 +77,20 @@ def ecef2aer(x, y, z, lat0, lon0, h0, ell=None, deg=True):
 
 def eci2aer(eci, lat0, lon0, h0, t):
     """
-    Convert ECI of target to azimuth, elevation, range (meters) from observer at lat0,lon0,h0 (meters)
-    output: azimuth (deg), elevation (deg), slant range (m) for Obs->Point
+    Observer => Point
+
+    input
+    -----
+    eci [meters] Nx3 target ECI location (x,y,z)                    [0,Infinity)
+    lat0, lon0 (degrees/radians)  Observer coordinates on ellipsoid [-90,90],[-180,180]
+    h0     [meters]                observer altitude                [0,Infinity)
+    t  time (datetime.datetime)   time of obsevation (UTC)
+
+
+    output: AER
+    ------
+    azimuth, elevation (degrees/radians)                             [0,360),[0,90]
+    slant range [meters]                                             [0,Infinity)
     """
     ecef = eci2ecef(eci, t)
 
@@ -75,11 +99,17 @@ def eci2aer(eci, lat0, lon0, h0, t):
 
 def enu2aer(e, n, u, deg=True):
     """
-    Convert ENU (meters) to azimuth, elevation, range (meters)
+    Observer => Point
 
-    input: east, north, up [m]
+    input
+    -----
+    e,n,u  [meters]  East, north, up                                [0,Infinity)
+    deg    degrees input/output  (False: radians in/out)
 
-    output: azimuth (deg), elevation (deg), slant range (m) for Obs->Point
+    output: AER
+    ------
+    azimuth, elevation (degrees/radians)                             [0,360),[0,90]
+    slant range [meters]                                             [0,Infinity)
     """
     r = hypot(e, n)
     slantRange = hypot(r, u)
@@ -93,12 +123,19 @@ def enu2aer(e, n, u, deg=True):
 
 def geodetic2aer(lat, lon, h, lat0, lon0, h0, ell=None, deg=True):
     """
[…]
