+++ CMakeLists.txt
@@ -5,6 +5,11 @@ add_compile_options(-ffpe-trap=invalid,zero,overflow -g -fbacktrace
                     -Wall -Wextra -Warray-bounds)# -Wfatal-errors)
 
 add_library(maptran fortran/maptran.f90 fortran/vallado.f90 fortran/assert.f90 )
+# --- testing
+enable_testing()
 
 add_executable(testmaptran fortran/Test.f90)
 target_link_libraries(testmaptran maptran)
+
+add_test(NAME testmaptran COMMAND testmaptran)
+
