### 2018-03-12 scivision/pymap3d#12 @scivision
Thanks, I noted how to run the tests from each language.
