### 2018-03-12 scivision/pymap3d#13 @scivision
Thanks  I moved it to the top level of the repo
