### 2018-03-12 #580 @lheagy
@arfon: this submission is ready for publication in JOSS.

Congratulations @scivision! Thanks @hugoledoux and @leouieda for reviewing!
